requirements.txt
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
dnacentersdk==2.4.9
fastjsonschema==2.15.3
future==0.18.2
idna==3.3
O365==2.0.18.1
oauthlib==3.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
six==1.16.0
soupsieve==2.3.2.post1
stringcase==1.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
